+++ packages/demos/browser/vite.config.ts
@@ -1,7 +1,8 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import cesiumEngine from "vite-plugin-cesium-engine";
+import { cesiumEngine } from "vite-plugin-cesium-engine";
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
+  base: mode === "pages" ? "/acmi-parser/" : "/",
   plugins: [react(), cesiumEngine()],
-});
+}));